+++ src/main/java/OMOP/v54/DrugExposure.java
@@ -1,5 +1,6 @@
 package OMOP.v54;
 
+import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 import java.util.List;
 import java.util.Optional;
@@ -176,9 +177,9 @@ public class DrugExposure {
   }
   
   @Column(name = "quantity", insertable = false, updatable = false)
-  private Float quantity;
+  private BigDecimal quantity;
   
-  public Optional<Float> getQuantity() {
+  public Optional<BigDecimal> getQuantity() {
     if (this.quantity != null) {
       return Optional.of(this.quantity);
     } else {
@@ -329,12 +330,13 @@ public class DrugExposure {
     }
   }
   
-@Override
-public String toString() {
-    final var result = new StringBuilder();
-    result.append("DrugExposure{id=").append(this.drugExposureId);
-    result.append("}");
-    return result.toString();
-}
+  
+  @Override
+  public String toString() {
+      final var result = new StringBuilder();
+      result.append("DrugExposure{id=").append(this.drugExposureId);
+      result.append("}");
+      return result.toString();
+  }
   
 }
