+++ src/main/java/OMOP/v54/Observation.java
@@ -1,5 +1,6 @@
 package OMOP.v54;
 
+import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 import java.util.List;
 import java.util.Optional;
@@ -230,9 +231,9 @@ public class Observation {
   }
   
   @Column(name = "value_as_number", insertable = false, updatable = false)
-  private Float valueAsNumber;
+  private BigDecimal valueAsNumber;
   
-  public Optional<Float> getValueAsNumber() {
+  public Optional<BigDecimal> getValueAsNumber() {
     if (this.valueAsNumber != null) {
       return Optional.of(this.valueAsNumber);
     } else {
@@ -328,17 +329,18 @@ public class Observation {
     }
   }
   
-@Override
-public String toString() {
-    final var result = new StringBuilder();
-    result.append("Observation{id=").append(this.observationId);
-    this.getObservationConcept().ifPresent(concept -> {
-      result.append(", concept='")
-      .append(concept.getConceptName().get())
-      .append("'");
-    });
-    result.append("}");
-    return result.toString();
-}
+  
+  @Override
+  public String toString() {
+      final var result = new StringBuilder();
+      result.append("Observation{id=").append(this.observationId);
+      this.getObservationConcept().ifPresent(concept -> {
+        result.append(", concept='")
+        .append(concept.getConceptName().get())
+        .append("'");
+      });
+      result.append("}");
+      return result.toString();
+  }
   
 }
